Added basic component definition generation
