+++ main.go
@@ -4,7 +4,7 @@ Copyright © 2025 NAME HERE <endome44@protonmail.com>
 */
 package main
 
-import "github.com/open-automation-construct/stigctl/cmd"
+import "github.com/open-automation-construct/stigctl/src/cmd"
 
 func main() {
 	cmd.Execute()
